Completed Project

+++ category.php
@@ -5,52 +5,55 @@
 
     if ($_POST)
     {
-        $filteredCategory  = filter_input(INPUT_POST, 'category', FILTER_SANITIZE_FULL_SPECIAL_CHARS);
-        $filteredId = filter_input(INPUT_POST, 'id', FILTER_SANITIZE_NUMBER_INT);
-
-        if ($_POST['post'] == "add")
-        {
-            $query = "INSERT INTO category (categoryName) VALUES (:categoryName)";
-    
-            $statement = $db->prepare($query);
-                
-            $statement->bindValue(":categoryName", $filteredCategory);
-            
-            // Execute the INSERT.
-            $statement->execute();
-    
-            header("Location: category.php");
-            exit;
+        if (isset($_POST['category']) && empty($_POST['category'])){
+           $categoryError = true; 
         }
-
-        if ($_POST['post'] == "delete")
-        {
-            $query = "DELETE FROM category WHERE categoryId = :categoryId LIMIT 1";
-
-            $statement = $db->prepare($query);
-                
-            $statement->bindValue(":categoryId", $filteredId, PDO::PARAM_INT);
-            
-            // Execute the INSERT.
-            $statement->execute();
+        else {
+            $filteredCategory  = filter_input(INPUT_POST, 'category', FILTER_SANITIZE_FULL_SPECIAL_CHARS);
+            $filteredId = filter_input(INPUT_POST, 'id', FILTER_SANITIZE_NUMBER_INT);
     
-            header("Location: category.php");
-            exit;
-        }
-
-        if ($_POST['post'] == "update")
-        {
-            $query = "UPDATE category SET categoryName = :categoryName WHERE categoryId = :categoryId";
-            
-            $statement = $db->prepare($query);
+            if ($_POST['post'] == "add") {
+                $query = "INSERT INTO category (categoryName) VALUES (:categoryName)";
+        
+                $statement = $db->prepare($query);
+                    
+                $statement->bindValue(":categoryName", $filteredCategory);
                 
-            $statement->bindValue(":categoryId", $filteredId, PDO::PARAM_INT);
-            $statement->bindValue(":categoryName", $filteredCategory);
-
-            $statement->execute();
+                // Execute the INSERT.
+                $statement->execute();
+        
+                header("Location: category.php");
+                exit;
+            }
     
-            header("Location: category.php");
-            exit;
+            if ($_POST['post'] == "delete"){
+                $query = "DELETE FROM category WHERE categoryId = :categoryId LIMIT 1";
+    
+                $statement = $db->prepare($query);
+                    
+                $statement->bindValue(":categoryId", $filteredId, PDO::PARAM_INT);
+                
+                // Execute the INSERT.
+                $statement->execute();
+        
+                header("Location: category.php");
+                exit;
+            }
+    
+            if ($_POST['post'] == "update"){
+                
+                $query = "UPDATE category SET categoryName = :categoryName WHERE categoryId = :categoryId";
+                
+                $statement = $db->prepare($query);
+                    
+                $statement->bindValue(":categoryId", $filteredId, PDO::PARAM_INT);
+                $statement->bindValue(":categoryName", $filteredCategory);
+    
+                $statement->execute();
+        
+                header("Location: category.php");
+                exit;
+            }
         }
     }
     
@@ -81,12 +84,15 @@
                 <div class="bg-image" style="background-image: url('images/post.png');">
                     <h2 class="my-4 text-center fw-bolder">Categories List</h2>
                 </div>
+                <?php if (isset($categoryError) && $categoryError) : ?>
+                    <p class="text-white my-2 mx-2">Field cannot be empty.</p>
+                <?php endif ?>
                 <?php while ($category = $statement->fetch()) : ?>
                     <div class="categories">
                         <form action="category.php" method="post">
                             <?php if ($_POST && $_POST['post'] == 'edit ' . $category['categoryId']) : ?>
                                 <input type="hidden" name="id" value="<?= $category['categoryId'] ?>">
-                                <input type="text" name="category" value=<?= $category['categoryName'] ?> autofocus  onfocus="this.select()">
+                                <input type="text" name="category" value=<?= $category['categoryName'] ?> autofocus onfocus="this.select()">
                                 <button class="btn btn-primary btn-sm my-2 mx-2" type="submit" name="post" value="update">Update</button>
                             <?php else : ?>
                                 <input type="hidden" name="id" value="<?= $category['categoryId'] ?>">
@@ -100,7 +106,7 @@
                     </div>
                 <?php endwhile ?>
                 <form action="category.php" method="post">
-                    <label class="my-2 mx-2 text-white" for="category">Add a category</label>
+                    <label class="my-4 mx-2 text-white" for="category">Add a category</label>
                     <input type="text" name="category" id="category">
                     <button class="btn btn-success btn-sm my-2 mx-1" type="submit" name="post" value="add">Add</button>
                 </form>
